tools/content_analysis_tools.py
"""Content analysis tools for the ContentAnalyzer agent."""

from strands import tool
from typing import Dict, List, Any


@tool
def analyze_content_structure(content: str) -> Dict[str, Any]:
    """Analyze text content to extract key points and structure.
    
    Args:
        content: The input text content to analyze
        
    Returns:
        Dictionary containing analyzed content structure
    """
    # TODO: Implement content analysis logic
    return {
        "key_points": [],
        "structure": {},
        "topics": [],
        "sentiment": "neutral"
    }


@tool
def extract_key_messages(content: str, max_points: int = 5) -> List[str]:
    """Extract the most important messages from content.
    
    Args:
        content: The input text content
        max_points: Maximum number of key points to extract
        
    Returns:
        List of key message strings
    """
    # TODO: Implement key message extraction
    return []


@tool
def categorize_content_type(content: str) -> str:
    """Determine the type/category of the content.
    
    Args:
        content: The input text content
        
    Returns:
        Content category string
    """
    # TODO: Implement content categorization
    return "general"


def get_content_analysis_tools():
    """Return list of content analysis tools for agent initialization."""
    return [
        analyze_content_structure,
        extract_key_messages,
        categorize_content_type
    ]