pyproject.toml
[project]
name = "aws-infographic-generator"
version = "0.1.0"
description = "AWS-native multi-agent infographic generator using Amazon Bedrock, S3, and Nova Canvas for the AWS AI Agent Global Hackathon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-requests-auth>=0.4.3",
    "boto3>=1.38.36",
    "botocore>=1.35.36",
    "pillow>=11.2.1",
    "pydantic>=2.0.0",
    "requests>=2.32.4",
    "strands-agents>=1.0.1",
    "strands-agents-tools>=0.2.2",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true